requirements.txt
imdbpy
joblib
requests
scikit-learn
streamlit
watchdog